fix(live): explain an empty feed where the reader actually meets it
#3518 only reached feeds with no events at all. But a started agent reports
its lifecycle before its CLI starts refusing turns, so the feed that gets
reported as broken does have events — just no tool events — and it took the
drill-down's "No tool events yet for this agent" path instead. True,
unhelpful, and indistinguishable from a slow start. That line is what "I
only see state changed events" was looking at.

EmptyFeed moves into its own module and the drill-down accepts it as an
optional empty state, so both paths give the same explanation. Verified
against a real stuck agent — pi on a model it has no access to, running 31
days — which now reads "Nothing reported since this agent started 31d ago"
and offers the terminal that names the reason.

Two smaller corrections to that copy, both found by looking at it:
formatRelative's default fallback to an absolute date past 30 days turned
the sentence into "started 2 Jul 2026", and the detail line repeated the
headline's claim that nothing is being reported.
