+++ web/src/components/live/AgentToolStream.tsx
@@ -1,8 +1,7 @@
-import { useEffect, useState, type ReactNode } from "react";
-import { Link } from "react-router-dom";
+import { useEffect, useState } from "react";
 import { api } from "../../api/client";
 import { useAgentActivity } from "../../hooks/useAgentActivity";
-import { formatRelative } from "../../utils/time";
+import { EmptyFeed } from "./EmptyFeed";
 import { AgentDrillDown } from "./LiveRenderers";
 
 /* ── AgentToolStream ────────────────────────────────────────────────
@@ -66,105 +65,6 @@ function useReportsActivity(agentTool?: string): boolean | undefined {
   return mode === undefined || mode !== "none";
 }
 
-/**
- * How long a running agent may report nothing before the empty feed stops
- * saying "waiting" and starts saying "something is wrong".
- *
- * Every provider mycel can capture emits something within seconds of starting
- * — a session-start hook, or the first transcript line. Minutes of total
- * silence from a running agent is not a slow start, it is a broken one.
- */
-const SILENCE_IS_SUSPICIOUS_AFTER_MS = 2 * 60 * 1000;
-
-/**
- * Re-renders on an interval so a feed that never receives a single event can
- * still notice that time has passed. Without this, the "waiting for the first
- * event" copy would be frozen forever on exactly the agents that most need to
- * be told something is wrong, because no event ever arrives to re-render it.
- */
-function useElapsingTime(active: boolean, everyMs = 30_000): void {
-  const [, setTick] = useState(0);
-  useEffect(() => {
-    if (!active) return;
-    const id = setInterval(() => setTick((n) => n + 1), everyMs);
-    return () => clearInterval(id);
-  }, [active, everyMs]);
-}
-
-/**
- * What to say when there is nothing to show.
- *
- * An empty feed has several very different causes and they are not
- * interchangeable: a provider mycel cannot observe, an agent that is not
- * running, an agent that just started, and an agent that is running but whose
- * CLI is refusing every turn. The last one is the one that gets reported as
- * "live doesn't work" — the events are missing because the agent is dead, and
- * the only place that says so is its terminal.
- */
-function EmptyFeed({
-  agentName,
-  agentTool,
-  agentState,
-  startedAt,
-  reportsActivity,
-}: {
-  agentName: string;
-  agentTool?: string;
-  agentState: string;
-  startedAt?: string;
-  reportsActivity: boolean | undefined;
-}) {
-  const running = agentState !== "stopped" && agentState !== "error";
-  useElapsingTime(running);
-
-  const startedMs = startedAt ? new Date(startedAt).getTime() : NaN;
-  const silentTooLong =
-    running &&
-    Number.isFinite(startedMs) &&
-    Date.now() - startedMs > SILENCE_IS_SUSPICIOUS_AFTER_MS;
-
-  let headline: ReactNode;
-  let detail: ReactNode;
-  let attachLabel: string | null = null;
-
-  if (reportsActivity === false) {
-    headline = <>mycel can&rsquo;t read activity from {agentTool} agents</>;
-    detail =
-      "This provider offers neither hooks nor a transcript to read, so there is nothing to stream. Its terminal is the whole picture.";
-    attachLabel = "Watch the terminal";
-  } else if (!running) {
-    headline = "This agent isn't running";
-    detail =
-      "Start it from the header and its prompts, tool calls, and results will appear here.";
-  } else if (silentTooLong && reportsActivity !== undefined) {
-    headline = <>Nothing reported since this agent started {formatRelative(startedAt)}</>;
-    detail =
-      "An agent that can't reach its model — missing key, spent quota, a model it isn't entitled to — runs but never works, and reports nothing here. Its terminal will name the reason.";
-    attachLabel = "See what the terminal says";
-  } else {
-    headline = "Waiting for the first event";
-    detail =
-      "Prompts, tool calls, and results appear here as the agent works.";
-  }
-
-  return (
-    <div className="flex-1 flex items-center justify-center p-6">
-      <div className="max-w-sm flex flex-col gap-2 text-center">
-        <p className="text-sm font-medium text-mycel-text">{headline}</p>
-        <p className="text-[13px] text-mycel-muted leading-relaxed">{detail}</p>
-        {attachLabel && (
-          <Link
-            to={`/agents/${agentName}/attach`}
-            className="mt-1 self-center text-[12px] text-mycel-accent rounded px-2 py-1 ring-1 ring-inset ring-mycel-border hover:bg-mycel-surface-hover focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-mycel-accent"
-          >
-            {attachLabel}
-          </Link>
-        )}
-      </div>
-    </div>
-  );
-}
-
 export function AgentToolStream({
   agentName,
   agentTool,
@@ -177,17 +77,21 @@ export function AgentToolStream({
   const activity = activities.get(agentName);
   const rawEvents = rawEventsRef.current.get(agentName) ?? [];
 
-  if (!activity) {
-    return (
-      <EmptyFeed
-        agentName={agentName}
-        agentTool={agentTool}
-        agentState={agentState}
-        startedAt={startedAt}
-        reportsActivity={reportsActivity}
-      />
-    );
-  }
+  // Same explanation whether nothing at all arrived or only lifecycle events
+  // did. Both are the same situation to the reader — an empty feed — and the
+  // second is the more common one, since a started agent reports that much even
+  // when its CLI then refuses every turn.
+  const emptyFeed = (
+    <EmptyFeed
+      agentName={agentName}
+      agentTool={agentTool}
+      agentState={agentState}
+      startedAt={startedAt}
+      reportsActivity={reportsActivity}
+    />
+  );
+
+  if (!activity) return emptyFeed;
 
   return (
     <div className="p-6 flex flex-col h-full">
@@ -196,6 +100,7 @@ export function AgentToolStream({
         rawEvents={rawEvents}
         tasks={tasks}
         hideRawStream
+        emptyState={emptyFeed}
       />
     </div>
   );
